Consolidate producer header tests and fix StreamDB live test for fetch-then-live
Move producer header merge tests into headers-params-functions.test.ts,
remove accidental .opentrace/.gitignore, and update StreamDB live mode
test to account for the fetch-then-live pattern where the initial request
omits the live param for CDN cacheability.
