--- .opentrace/.gitignore
@@ -1,4 +1,0 @@
-# OpenTrace index data — generated by `opentraceai index`
-*.db
-*.db.wal
-*.indexlock
